move splitPivotData logic into multiLevelPivot
